+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/aio/_async_container_registry_client.py
@@ -33,13 +33,14 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param endpoint: An ACR endpoint
         :type endpoint: str
         :param credential: The credential with which to authenticate
-        :type credential: :class:`~azure.core.credentials_async.AsyncTokenCredential`
+        :type credential: ~azure.core.credentials_async.AsyncTokenCredential
         :keyword audience: URL to use for credential authentication with AAD. Its value could be
         "https://management.azure.com", "https://management.chinacloudapi.cn", "https://management.microsoftazure.de" or
         "https://management.usgovcloudapi.net"
         :paramtype audience: ~azure.containerregistry.ContainerRegistryAudience or str
         :returns: None
-        :raises: None
+        :rtype: None
+        :raises: if the provided api_version keyword-only argument isn't supported
 
         .. admonition:: Example:
 
@@ -78,7 +79,7 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param str repository: The repository to delete
         :returns: None
         :rtype: None
-        :raises: :class:`~azure.core.exceptions.HttpResponseError`
+        :raises: ~azure.core.exceptions.HttpResponseError
 
         .. admonition:: Example:
 
@@ -99,7 +100,7 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :paramtype results_per_page: int
         :returns: An iterable of strings
         :rtype: ~azure.core.async_paging.AsyncItemPaged[str]
-        :raises: :class:`~azure.core.exceptions.HttpResponseError`
+        :raises: ~azure.core.exceptions.HttpResponseError
 
         .. admonition:: Example:
 
@@ -204,8 +205,8 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         """Get the properties of a repository
 
         :param str repository: Name of the repository
-        :returns: :class:`~azure.containerregistry.RepositoryProperties`
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :rtype: ~azure.containerregistry.RepositoryProperties
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
         """
         return RepositoryProperties._from_generated(  # pylint: disable=protected-access
             await self._client.container_registry.get_properties(repository, **kwargs)
@@ -217,12 +218,12 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
 
         :param str repository: Name of the repository
         :keyword order_by: Query parameter for ordering by time ascending or descending
-        :paramtype order_by: :class:`~azure.containerregistry.ManifestOrder` or str
+        :paramtype order_by: ~azure.containerregistry.ArtifactManifestOrder or str
         :keyword results_per_page: Number of repositories to return per page
         :paramtype results_per_page: int
         :returns: An iterable of :class:`~azure.containerregistry.ArtifactManifestProperties`
         :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.containerregistry.ArtifactManifestProperties]
-        :raises: :class:`~azure.core.exceptions.HttpResponseError`
+        :raises: ~azure.core.exceptions.HttpResponseError
         """
         name = repository
         last = kwargs.pop("last", None)
@@ -338,16 +339,16 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param str tag_or_digest: Tag or digest of the manifest to be deleted.
         :returns: None
         :rtype: None
-        :raises: :class:`~azure.core.exceptions.HttpResponseError`
+        :raises: ~azure.core.exceptions.HttpResponseError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, DefaultAzureCredential())
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
             await client.delete_manifest("my_repository", "my_tag_or_digest")
         """
         if _is_tag(tag_or_digest):
@@ -364,17 +365,17 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param str tag: The tag to be deleted
         :returns: None
         :rtype: None
-        :raises: :class:`~azure.core.exceptions.HttpResponseError`
+        :raises: ~azure.core.exceptions.HttpResponseError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, "my_repository", DefaultAzureCredential())
-            async for artifact in client.list_tag_properties():
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
+            async for tag in client.list_tag_properties("my_repository"):
                 await client.delete_tag("my_repository", tag.name)
         """
         await self._client.container_registry.delete_tag(repository, tag, **kwargs)
@@ -387,18 +388,18 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
 
         :param str repository: Name of the repository
         :param str tag_or_digest: The tag or digest of the manifest
-        :returns: :class:`~azure.containerregistry.ArtifactManifestProperties`
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :rtype: ~azure.containerregistry.ArtifactManifestProperties
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, "my_repository", DefaultAzureCredential())
-            async for artifact in client.list_manifest_properties():
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
+            async for artifact in client.list_manifest_properties("my_repository"):
                 properties = await client.get_manifest_properties("my_repository", artifact.digest)
         """
         if _is_tag(tag_or_digest):
@@ -417,18 +418,18 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param str repository: Repository the tag belongs to
         :param tag: The tag to get properties for
         :type tag: str
-        :returns: :class:`~azure.containerregistry.ArtifactTagProperties`
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :rtype: ~azure.containerregistry.ArtifactTagProperties
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, "my_repository", DefaultAzureCredential())
-            async for tag in client.list_tag_properties():
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
+            async for tag in client.list_tag_properties("my_repository"):
                 tag_properties = await client.get_tag_properties("my_repository", tag.name)
         """
         return ArtifactTagProperties._from_generated(  # pylint: disable=protected-access
@@ -442,22 +443,22 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
 
         :param str repository: Name of the repository
         :keyword order_by: Query parameter for ordering by time ascending or descending
-        :paramtype order_by: :class:`~azure.containerregistry.TagOrder` or str
+        :paramtype order_by: ~azure.containerregistry.ArtifactTagOrder or str
         :keyword results_per_page: Number of repositories to return per page
         :paramtype results_per_page: int
         :returns: An iterable of :class:`~azure.containerregistry.ArtifactTagProperties`
         :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.containerregistry.ArtifactTagProperties]
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, "my_repository", DefaultAzureCredential())
-            async for tag in client.list_tag_properties():
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
+            async for tag in client.list_tag_properties("my_repository"):
                 tag_properties = await client.get_tag_properties(tag.name)
         """
         name = repository
@@ -586,9 +587,9 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :type args: Union[str, ~azure.containerregistry.RepositoryProperties]
         :param str repository: Name of the repository
         :param properties: Properties to set for the repository
-        :type properties: :class:`~azure.containerregistry.RepositoryProperties`
-        :returns: :class:`~azure.containerregistry.RepositoryProperties`
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :type properties: ~azure.containerregistry.RepositoryProperties
+        :rtype: ~azure.containerregistry.RepositoryProperties
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
         """
         repository, properties = None, None
         if len(args) == 2:
@@ -633,18 +634,18 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param str tag_or_digest: Tag or digest of the manifest
         :param properties: The property's values to be set
         :type properties: ArtifactManifestProperties
-        :returns: :class:`~azure.containerregistry.ArtifactManifestProperties`
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :rtype: ~azure.containerregistry.ArtifactManifestProperties
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, "my_repository", DefaultAzureCredential())
-            async for artifact in client.list_manifest_properties():
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
+            async for artifact in client.list_manifest_properties("my_repository"):
                 received_properties = await client.update_manifest_properties(
                     "my_repository",
                     artifact.digest,
@@ -703,17 +704,17 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :param str tag: Tag to set properties for
         :param properties: The property's values to be set
         :type properties: ArtifactTagProperties
-        :returns: :class:`~azure.containerregistry.ArtifactTagProperties`
-        :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
+        :rtype: ~azure.containerregistry.ArtifactTagProperties
+        :raises: ~azure.core.exceptions.ResourceNotFoundError
 
         Example
 
         .. code-block:: python
 
-            from azure.containerregistry.aio import ContainerRepositoryClient
+            from azure.containerregistry.aio import ContainerRegistryClient
             from azure.identity.aio import DefaultAzureCredential
             endpoint = os.environ["CONTAINERREGISTRY_ENDPOINT"]
-            client = ContainerRepositoryClient(endpoint, "my_repository", DefaultAzureCredential())
+            client = ContainerRegistryClient(endpoint, DefaultAzureCredential(), audience="my_audience")
             tag_identifier = "latest"
             received = await client.update_tag_properties(
                 "my_repository",